+++ src/js/controller/CarManager.js
@@ -1,9 +1,9 @@
-import Car from "../../model/Car.js";
-import { isValidCarsName } from "./validation.js";
-import { showRacingCountArea } from "../../view/viewControl.js";
-import { EXCEPTIONS } from "../../constants/exceptions.js";
-import { $ } from "../../view/getElement.js";
-import { trimArray } from "../../utils/trimArray.js";
+import Car from "../model/Car.js";
+import { isValidCarsName } from "./carManager/validation.js";
+import { showRacingCountArea } from "../view/viewControl.js";
+import { EXCEPTIONS } from "../constants/exceptions.js";
+import { $ } from "../view/getElement.js";
+import { trimArray } from "../utils/trimArray.js";
 
 // TODO : 자동차들을 만들고 관리
 export default class CarManager {
